+++ src/block/class.ts
@@ -19,7 +19,7 @@ export default class Class extends Block
     public parse():Doc
     {
         let params = this.match();
-        let doc = new Doc(TypeUtil.instance.getDefaultMessage(params[3], params[2]));
+        let doc = new Doc(Doc.TYPE_CLASS, TypeUtil.instance.getDefaultMessage(params[3], params[2]));
         doc.template = Config.instance.get('classTemplate');
 
         return doc;